[#50] rename: domain 폴더로 이동 후 import 문 수정
