+++ auth/src/main/java/com/devcamp/flametalk/domain/user/dto/GatewayUserDto.java
@@ -1,4 +1,4 @@
-package com.devcamp.flametalk.user.dto;
+package com.devcamp.flametalk.domain.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
