+++ auth/src/main/java/com/devcamp/flametalk/global/util/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package com.devcamp.flametalk.global.util;
 
-import com.devcamp.flametalk.token.service.TokenService;
+import com.devcamp.flametalk.domain.token.service.TokenService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.Jwts;
@@ -17,7 +17,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
